requirements.txt
gradio==5.25.2
gradio-rangeslider==0.0.8
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
psutil==7.0.0
tqdm==4.67.1
scipy==1.15.2
# RunPod SDK
runpod
# Storage dependencies (optional)
google-cloud-storage
boto3
requests
